initializes dependencies up front in main
the logger will be initialized and a connection to the db will be
opened once, up front, in main. then a reference to the logger and
db is passed through to all of the handlers. this provides a handful
of benefits:
- the handlers don't have to be concerned with dependency initialization
  (separation of concerns)
- any errors during initialization of these dependencies will happen on
  boot, instead of potentially during request handling (fail loud, fail
  fast)

this PR also cleans up some error handling throughout the various
utilities and handlers